requirements.txt
matplotlib==3.7.1
numpy==1.24.1
opencv_python==4.7.0.72
pandas==2.0.1
Pillow==9.5.0
scikit_learn==1.2.2
six==1.16.0
torch
torchvision
tqdm==4.23.4
